requirements.txt
openai==2.9.0
python-dotenv==1.2.1
Requests==2.32.5
streamlit==1.51.0
streamlit_autorefresh==1.0.1
